refactor!: Removes dependencies from host on provider libraries
Our current host had a circular dependency loop with itself because it
depended on the http and messaging providers. In order to break this, I
moved the common http and messaging types we use for builtins into
`wasmcloud-core` behind feature flags that aren't on by default. As such,
this is a breaking change as I moved stuff around, but mostly didn't
change any code.

Please note that I only bumped version numbers on things we had released
already. Some of these crates had not yet been released and do not need
another bump
