+++ crates/provider-http-server/Cargo.toml
@@ -31,6 +31,7 @@ tokio = { workspace = true, features = ["macros"] }
 tower-http = { workspace = true, features = ["cors"] }
 tracing = { workspace = true }
 unicase = { workspace = true }
+wasmcloud-core = { workspace = true, features = ["http"] }
 wasmcloud-provider-sdk = { workspace = true, features = ["otel"] }
 wrpc-interface-http = { workspace = true, features = ["http-body"] }
