+++ crates/provider-messaging-nats/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasmcloud-provider-messaging-nats"
-version = "0.26.0"
+version = "0.27.0"
 description = """
 A capability provider that satisfies the 'wasmcloud:messaging' contract using NATS as a backend.
 """
@@ -27,5 +27,6 @@ tokio = { workspace = true, features = ["full"] }
 tracing = { workspace = true }
 tracing-futures = { workspace = true }
 wascap = { workspace = true }
+wasmcloud-core = { workspace = true, features = ["messaging"] }
 wasmcloud-provider-sdk = { workspace = true, features = ["otel"] }
 wit-bindgen-wrpc = { workspace = true }
