+++ crates/provider-sdk/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasmcloud-provider-sdk"
-version = "0.14.0"
+version = "0.15.0"
 description = "wasmCloud provider SDK"
 readme = "README.md"
