+++ crates/runtime/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasmcloud-runtime"
-version = "0.9.0"
+version = "0.10.0"
 description = "wasmCloud runtime library"
 readme = "README.md"
