+++ crates/tracing/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "wasmcloud-tracing"
-version = "0.13.0"
+version = "0.14.0"
 description = "wasmCloud tracing functionality"
 
 authors.workspace = true
